+++ dart/geobase/test/geodesy/utm_test.dart
@@ -6,7 +6,6 @@
 
 // ignore_for_file: require_trailing_commas, lines_longer_than_80_chars, avoid_redundant_argument_values
 
-import 'package:geobase/common.dart';
 import 'package:geobase/coordinates.dart';
 import 'package:geobase/geodesy.dart';