+++ dart/geobase/test/projections/projection_ellipsoidal_test.dart
@@ -8,7 +8,6 @@
 
 import 'dart:math';
 
-import 'package:geobase/common.dart';
 import 'package:geobase/coordinates.dart';
 import 'package:geobase/geodesy.dart';
 import 'package:geobase/projections.dart';