+++ dart/geodata/test/geodata_core_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2024 Navibyte (https://navibyte.com). All rights reserved.
+// Copyright (c) 2020-2025 Navibyte (https://navibyte.com). All rights reserved.
 // Use of this source code is governed by a “BSD-3-Clause”-style license that is
 // specified in the LICENSE file.
 //
@@ -8,7 +8,6 @@
 // ignore_for_file: prefer_const_constructors,require_trailing_commas
 // ignore_for_file: prefer_const_literals_to_create_immutables
 
-import 'package:geobase/common.dart';
 import 'package:geobase/coordinates.dart';
 import 'package:geobase/meta.dart';
 import 'package:geodata/core.dart';